+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1162](https://github.com/alleslabs/celatone-frontend/pull/1162) Proxy move verify through celatone verification api first
+
 ### Bug fixes
 
 ## v1.9.1