+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1169](https://github.com/alleslabs/celatone-frontend/pull/1169) Fix minor bugs
 - [#1168](https://github.com/alleslabs/celatone-frontend/pull/1168) Normalized voting power by 1e6 for Initia
 - [#1162](https://github.com/alleslabs/celatone-frontend/pull/1162) Proxy move verify through celatone verification api first