+++ CHANGELOG.md
@@ -35,6 +35,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.9.2
+
+### Improvements
+
+- [#1169](https://github.com/alleslabs/celatone-frontend/pull/1169) Fix minor bugs
+- [#1168](https://github.com/alleslabs/celatone-frontend/pull/1168) Normalized voting power by 1e6 for Initia
+- [#1162](https://github.com/alleslabs/celatone-frontend/pull/1162) Proxy move verify through celatone verification api first
+
 ## v1.9.1
 
 ### Improvements